# ticketing/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the TicketPulse API for ticketing microservice. Concerts, sessions, sections and tickets are all here.",
        "title": "TicketPulse Ticketing API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/sessions": {
            "get": {
                "description": "Retrieves a list of all concert sessions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get all Concert Sessions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SessionListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new concert session with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a new Concert Session",
                "parameters": [
                    {
                        "description": "Date and time of the session",
                        "name": "sessionDateTime",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "ID of the concert",
                        "name": "concertId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/sessions/{id}": {
            "get": {
                "description": "Retrieves a concert session by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get a Concert Session by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the date and time of a concert session.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update a Concert Session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Date and time of the session",
                        "name": "sessionDateTime",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a concert session and its children by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a Concert Session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/sessions/{id}/sections": {
            "get": {
                "description": "Retrieves sections associated with a concert session by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get Sections by Session ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SessionWithSectionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ticketing_internal_common_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "ticketing_internal_common_types.Section": {
            "type": "object",
            "properties": {
                "bought": {
                    "type": "integer"
                },
                "capacity": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "reserved": {
                    "type": "integer"
                }
            }
        },
...
        "ticketing_internal_common_types.Session": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "sessionDateTime": {
                    "type": "integer"
                }
            }
        },
        "ticketing_internal_common_types.SessionListResponse": {
            "type": "object",
            "properties": {
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ticketing_internal_common_types.Session"
                    }
                }
            }
        },
        "ticketing_internal_common_types.SessionResponse": {
            "type": "object",
            "properties": {
                "session": {
                    "$ref": "#/definitions/ticketing_internal_common_types.Session"
                }
            }
        },
        "ticketing_internal_common_types.SessionWithSectionListResponse": {
            "type": "object",
            "properties": {
                "sections": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ticketing_internal_common_types.Section"
                    }
                },
                "session": {
                    "type": "string"
                }
            }
        },
        "ticketing_internal_common_types.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}